Drop arguments when the default is wanted
The modules line(), dimensions(), circle_center(), etc, already have
sensible defaults for the arguments. For internal users, there is no
need to specify those arguments explicitly when the default value really
is what is wanted. Remove the explicit arguments through most of the
library, except where keeping it in is better for readability.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>