Construct all dimensions as 2D objects.
Dimension lines are all effectively 2D drawings, but the code is
producing 3D objects that sometimes need to be projected into 2D before
they can be used for output. To remove the need for a projection step
(which is slow), restructure the code to always produce 2D objects and
add a wrapper module that extrudes the objects as needed depending on
the value of the $dim_extrude_flag special variable. $dim_extrude_flag
can be set by any file using the dimlines.scad file.

Extrude height is hard-coded because normally the height doesn't matter since
dimensions are viewed straight on. All that matters is whether they are 2D or
3D objects depending on the output format. SVG export will only work on 2D
objects. Image output only works for 3D objects. So, make it simply a flag that
defaults to extruding.

Anyone who needs them to be 2D can set $dim_extrude_flag=false;

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>