Add ability to change font
Add the $dim_font special variable to change the font used for
dimensions. By default it is not set, and the default font will be used.
If it is set in the caller code, then that value will be passed into all
instances of text().

This is also useful to deal with a minor font bug in the Debian packaged
version of OpenSCAD. The built-in copy of Liberation Sans doesn't
include the diameter symbol '⌀', which is useful for drawing leader
lines. Using FreeSans font instead solves the problem.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>