Remove dimension scaling from titleblock()
Auto scaling dimensions based on $dim_linewidth in titleblock() is a
little opaque, and it makes it hard to control how a title block will
actually appear. Remove all multiplication with dim_linewidth() from
titleblock() for determining line length. Instead, use the dimensions
exactly as given in the lines, descs and details arrays. However,
dim_linewidth() is still used to scale the line width argument in the
lines table, but that is because that parameter is already a relative
value intended to be scaled.

The undesirable side effect of this is the title block won't change size
if the font size is changed, which makes it harder to scale diagrams. To
mitigate this, change the recommended practice as shown in the sample
code to scale all table dimensions against the font size.  That way if
the font size is changed, all the text will still fit inside.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>