Add centering, leader lines, and auto-placement to dimensions()
Add two new arguments to dimensions(). 'offset' adjusts the placement of
the dimension line on the y-axis and draws leader lines on either side.
'center' will center the dimension line on the x axis.

Also add auto-placement of text labels. If the width is too small to fix
the text label, then place the text to the left.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>