Add $dim_modelscale to get consistent lines on scaled models
When making scaled drawings, the dimension lines get scaled along with
the object, which results in the scaled lines being much thinner than
the lines on non-scaled objects. ie. The title box font size will be
bigger than the dimension markings on a scaled model. When
$dim_modelscale is set, it tells dimlines how much the model is going to
be scaled down in the final picture, and it will scale up the line width
and font size accordingly so that everything looks correct.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>