Add support for different kinds of line endings.
In addition to arrows, add new "round" and "square" line endings.
"round" places a circle at the end of the line which fills in gaps when
butting up against another line. The leader line code is modified to use
this line ending.  "square" extends the length of the line by the
width/2. This is useful for joining square corners, and the titleblock
code is modified to use this feature. Finally, adjust the sample code to
show the line features.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>